Move PNG-rendering functionality into splash.qtrender_png module
Also update imports in related modules.
