+++ splash/browser_tab.py
@@ -1,28 +1,26 @@
 # -*- coding: utf-8 -*-
 from __future__ import absolute_import
-from cStringIO import StringIO
-import os
 import base64
 import copy
-import json
-import pprint
-import weakref
 import functools
-from PIL import Image
-from PyQt4.QtWebKit import QWebPage, QWebSettings, QWebView
-from PyQt4.QtCore import Qt, QUrl, QSize, QTimer, QObject, pyqtSlot
-from PyQt4.QtGui import QMouseEvent, QKeyEvent
+import json
+import os
+import weakref
+
+from PyQt4.QtCore import QObject, QSize, Qt, QTimer, QUrl, pyqtSlot
 from PyQt4.QtNetwork import QNetworkRequest
+from PyQt4.QtWebKit import QWebPage, QWebSettings, QWebView
 from twisted.internet import defer
 from twisted.python import log
+
 from splash import defaults
-from splash.qtutils import (qurl2ascii, OPERATION_QT_CONSTANTS, qt2py,
-                            WrappedSignal, render_qwebpage)
 from splash.har.qt import cookies2har
 from splash.har.utils import without_private
+from splash.qtrender_png import render_qwebpage
+from splash.qtutils import (OPERATION_QT_CONSTANTS, WrappedSignal, qt2py,
+                            qurl2ascii)
 from splash.render_options import validate_size_str
-
-from .qwebpage import SplashQWebPage
+from splash.qwebpage import SplashQWebPage
 
 
 def skip_if_closing(meth):
